lock_tables: introduce generic lock offset detection
Allow generating a table of stack offsets for fetching the lock from a
blocked thread's stack. This is done with a custom kernel module and an
exhaustive testing mechanism. The inputs and outputs are included in
this commit so we don't lose the information.

The resulting lock offsets are recorded in the README and will be used
in the next commit as a fallback for when DWARF is not available.

Signed-off-by: Stephen Brennan <stephen.s.brennan@oracle.com>