bt: fall back to the symbol name when indexing
We currently use the "frame.name" which may not be present when using
CTF. Instead, we should fallback to the symbol name.

The frame_name() function is not appropriate, since this may include
additional information, such as the module containing the symbol. So
introduce a new helper, func_name() for this case.

Signed-off-by: Stephen Brennan <stephen.s.brennan@oracle.com>