bt: handle "registers for CPU X were not saved"
Print an error message in bt(). Further, for the smp corelens module
which directly uses bt_frames(), add a handler for this bug as well.

This can happen for vmcores where a CPU did not respond to the interrupt
during the crash kexec operation. Drgn 0.0.27 recently learned to handle
this (see osandov/drgn@a47fa1f).

Signed-off-by: Stephen Brennan <stephen.s.brennan@oracle.com>