smp: make the output easier to scan by default
The smp module gives very useful information, but the current formatting
is a bit difficult to understand. The output is full of blank lines, and
there's no visual distinction between sections.

Adjust the output as follows. First, each CPU which has pending IPIs to
handle should get its own section. Use indentation to help visually
distinguish the different CPU source caller stack traces, and put the
messages related to each CPU source caller into one line. Remove
blank lines and rely on the indentation to provide visual grouping.

Finally, the major change is to remove the stack traces for each source
CPU (though it is retained as an option). In many cases, the source of
each IPI is not very important: what matters is that the IPI has not
been handled, and so the destination stack trace is what matters. What's
more, the same CPU can have a call pending for several destination CPUs,
meaning that its stack trace could be repeated several times over the
entire output. This added bulk makes it difficult to scan the
information quickly and understand the situation. By removing the source
stack traces by default, we gain a much cleaner view. For example,
here's an output on a recent issue I debugged:

    CPU: 76 has pending CSD requests:
      SRC 243 -> DST  76: waiting 70.648 secs, kworker/243:0:180493
      SRC 160 -> DST  76: waiting 83.859 secs, kworker/160:0:308566
      SRC   0 -> DST  76: **BUG:** Waiter CPU is idle! Waiter of a csd lock can't be idle.
      SRC   0 -> DST  76: **BUG:** Waiter CPU is idle! Waiter of a csd lock can't be idle.
    Call stack at destination cpu:  76
      PID: 0        TASK: ff31ebc8991ea080 [R] CPU: 76!  COMMAND: "swapper/76"
      error: registers for CPU 76 were not saved
    error: cannot determine whether IRQs are enabled on CPU 76: registers for CPU 76 were not saved
    destination is idle. CSD destination should not be idle when there are pending CSD requests.
    destination CPU's runq clock is lagging by 99.031 secs

    CPU: 268 has pending CSD requests:
      SRC 243 -> DST 268: waiting 70.648 secs, kworker/243:0:180493
      SRC 160 -> DST 268: waiting 83.859 secs, kworker/160:0:308566
    Call stack at destination cpu:  268
      PID: 0        TASK: ff31ebc8994bc100 [R] CPU: 268!  COMMAND: "swapper/268"
      error: registers for CPU 268 were not saved
    error: cannot determine whether IRQs are enabled on CPU 268: registers for CPU 268 were not saved
    destination is idle. CSD destination should not be idle when there are pending CSD requests.
    destination CPU's runq clock is lagging by 26.604 secs

Signed-off-by: Stephen Brennan <stephen.s.brennan@oracle.com>